move token to its module
