+++ packages/fuels-core/src/codec/logs.rs
@@ -31,9 +31,25 @@ impl LogFormatter {
         decoder_config: DecoderConfig,
         bytes: &[u8],
     ) -> Result<String> {
+        #[cfg(experimental)]
+        Self::can_decode_log_with_type::<T>()?;
+
         let token = ABIDecoder::new(decoder_config).decode(&T::param_type(), bytes)?;
 
         Ok(format!("{:?}", T::from_token(token)?))
+    }
+
+    #[cfg(experimental)]
+    fn can_decode_log_with_type<T: Parameterize>() -> Result<()> {
+        match T::param_type() {
+            // String slices cannot be decoded from logs as they are encoded as ptr, len
+            // TODO: Once https://github.com/FuelLabs/sway/issues/5110 is resolved we can remove this
+            ParamType::StringSlice => Err(error!(
+                Codec,
+                "string slices cannot be decoded from logs. Convert the slice to `str[N]` with `__to_str_array`"
+            )),
+            _ => Ok(()),
+        }
     }
 
     pub fn can_handle_type<T: Tokenizable + Parameterize + 'static>(&self) -> bool {
