+++ packages/fuels/tests/types/contracts/type_inside_enum/src/main.sw
@@ -3,7 +3,7 @@ contract;
 #[allow(dead_code)]
 enum SomeEnum {
     SomeStr: str[4],
-    SomeArr: [u64; 2],
+    SomeArr: [u64; 4],
 }
 
 #[allow(dead_code)]
